% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EloPackages.R
\name{current.day.prediction}
\alias{current.day.prediction}
\title{Build an Elo model}
\usage{
current.day.prediction(elo.table, date = Sys.Date())
}
\arguments{
\item{elo.table}{An Elo table generated using elo.table.generator}

\item{date}{The date through which the Elo ratings should be calculated. Should use YYYY-mm-dd format. Defaults to current date.}
}
\value{
A list of predictions for the given date.
}
\description{
This function generates predictions using an Elo table for the slate of games on a given date.
}
