% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EloPackages.R
\name{elo.by.team}
\alias{elo.by.team}
\title{Get team Elo ratings}
\usage{
elo.by.team(elo.model)
}
\arguments{
\item{elo.model}{An Elo model, which should be generated via elo.model.builder}
}
\value{
A dataframe of Elo ratings by team.
}
\description{
This function extracts Elo ratings for individual teams from an Elo model
}
