% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EloPackages.R
\name{elo.model.builder}
\alias{elo.model.builder}
\title{Build an Elo model}
\usage{
elo.model.builder(year, throughDate = Sys.Date())
}
\arguments{
\item{year}{The year for which the model should be built. Defaults to "back half" of years; for 2021-22, you would use year = 2022.}

\item{throughDate}{The date through which the Elo ratings should be calculated. Should use YYYY-mm-dd format.}
}
\value{
An Elo model for the specified time parameters.
}
\description{
This function generates an NBA Elo model for the given time parameters.
}
