% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EloPackages.R
\name{elo.prediction}
\alias{elo.prediction}
\title{Elo calculator}
\usage{
elo.prediction(elorating1, elorating2)
}
\arguments{
\item{elorating1}{Elo rating of Team 1.}

\item{elorating2}{Elo rating of Team 2.}
}
\value{
The probability of Team 1 beating Team 2.
}
\description{
This function computes the probability of Team 1 beating Team 2, given each team's Elo rating.
}
