% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EloPackages.R
\name{elo.standings.generator}
\alias{elo.standings.generator}
\title{Build a visualizable Elo rating table sorted by conference and division.}
\usage{
elo.standings.generator(overall.table)
}
\arguments{
\item{overall.table}{A table built using overall.table}
}
\value{
A GT table of overall ranks.
}
\description{
This function stylizes an input table, turning it into a HTML table using gt.
}
