% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EloPackages.R
\name{overall.function.wrapper}
\alias{overall.function.wrapper}
\title{Return all sorts of models}
\usage{
overall.function.wrapper(year, date = Sys.Date())
}
\arguments{
\item{year}{The year for which the model should be built. Defaults to "back half" of years; for 2021-22, you would use year = 2022.}

\item{date}{The date through which the Elo ratings should be calculated. Should use YYYY-mm-dd format.}
}
\value{
A list of tons of info for the given parameters.
}
\description{
This function generates an NBA Elo model, GT tables, etc. for the given time parameters.
}
