% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EloPackages.R
\name{overall.table}
\alias{overall.table}
\title{Combine an Elo table and an information table}
\usage{
overall.table(elo.table, info.table)
}
\arguments{
\item{elo.table}{The Elo table generated by elo.by.team}

\item{info.table}{The team info table generated via team.info.table}
}
\value{
A merged table.
}
\description{
This function combines the outputs of the elo.by.team and team.info.table functions into a single output.
}
