% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EloPackages.R
\name{team.info.table}
\alias{team.info.table}
\title{Get information on each team}
\usage{
team.info.table(year)
}
\arguments{
\item{year}{The year for which data should be collected.}
}
\value{
A table with information on team records, divisions, conferences, etc.
}
\description{
This function generates generic info about each NBA team for a given year.
}
