% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EloGeneration.R
\name{elo.by.team}
\alias{elo.by.team}
\title{Get Elo ratings by team}
\usage{
elo.by.team(elo.model)
}
\arguments{
\item{elo.model}{A valid Elo model generated using elo.model.builder}
}
\value{
A list of Elo ratings for each NBA team
}
\description{
Gives Elo ratings by team.
}
